moved fixtures away from tests
also pep8ted bap.utils.run module
